Check that exports succeeded
The `exports->Set` call here returns a `Maybe<bool>` indicating
whether it succeeded.  Before this change, we ignored this result.
This effectively never fails, but the unused result triggers a warning
in some compilers.

This change adds a Check() to this site to assert that the export
succeeded.  Note that this crashes the process on failure, but as
noted this failure shouldn't happen, and it's unlikely that clients
can usefully recover in any case.